ml-api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
redis>=5.0.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.45.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
supabase>=2.0.0
lxml>=5.0.0
optuna>=3.5.0
joblib>=1.3.0
aiohttp>=3.9.0
apscheduler>=3.10.0
pyarrow>=15.0.0
PublicDataReader>=0.4.0
